+++ assets/javascripts/kitten/karl/footer/footer-kisskiss.js
@@ -231,6 +231,8 @@ export class KarlFooterKisskiss extends Component {
                   margin={ false }
                   items={ items1 }
                   color="light"
+                  itemMargin="double"
+                  lineHeight="normal"
                 />
               </GridCol>
 
@@ -242,6 +244,8 @@ export class KarlFooterKisskiss extends Component {
                   margin={ false }
                   items={ items2 }
                   color="light"
+                  itemMargin="double"
+                  lineHeight="normal"
                 />
               </GridCol>
 
@@ -253,6 +257,8 @@ export class KarlFooterKisskiss extends Component {
                   margin={ false }
                   items={ items3 }
                   color="light"
+                  itemMargin="double"
+                  lineHeight="normal"
                 />
               </GridCol>
             </Grid>
@@ -593,11 +599,6 @@ KarlFooterKisskiss.defaultProps = {
     {
       key: 'key4',
       item: 'Les StaKissTiques',
-      href: '#',
-    },
-    {
-      key: 'key5',
-      item: 'KissKiss Cinéma',
       href: '#',
     },
   ],