+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `lineHeight` and `itemMargin` props on `LinkList` component.
+
 ## [19.4.1] - 2018-02-07
 
 Fix: